refactor components out of KubeSummaryPanel
